Initial components of user generated reference support: We encode user embeddings with pq against the scaffold of the current model's centroids. Resolve some sourcemap issues due to configuration. Various iterations on the figures notebook.

+++ ivfpq/python/src/ivfpq/pq.py
@@ -105,7 +105,8 @@ class PQ(nn.Module):
             # Find nearest centroid indices
             codes[:, i] = torch.argmin(distances, dim=1)
 
-        return codes
+        # Cast to uint8 for ONNX export (we always use 256 centroids max)
+        return codes.to(torch.uint8)
 
     def decode(self, codes: torch.Tensor) -> torch.Tensor:
         """
